Infer KV page count from HBM utilization (#1238)
## Summary
- add an `hbm_utilization` knob to `InferenceEngineConfig` and expose it
from the factory constructor
- infer a default KV page count via `eval_shape`/`tree_size` using the
device HBM budget when `max_pages` is omitted
- fall back to the previous `max_seqs * max_pages_per_seq` sizing when
memory statistics are unavailable

## Testing
- pre-commit run (via git commit hook)

fixes #1200 


------
https://chatgpt.com/codex/tasks/task_e_68e4b1ad30108331bee80d29e5b0195e

+++ src/levanter/layers/attention.py
@@ -1500,10 +1500,6 @@ class AttentionConfig:
             return default_attention_type() != AttentionBackend.VANILLA
         return self.attn_backend != AttentionBackend.VANILLA
 
-    # ---------------------------------------------------------------------------------
-    # KV-cache helper (paged only)
-    # ---------------------------------------------------------------------------------
-
 
 class Attention(eqx.Module):
     """A multi-head attention layer that uses dot product attention.
@@ -1700,8 +1696,8 @@ class Attention(eqx.Module):
         if self.rot_embs is not None:
             if pos_ids is None:
                 pos_ids = hax.arange(x.resolve_axis("position"))
-            q = self.rot_embs(q, pos_ids)
-            k = self.rot_embs(k, pos_ids)
+            q = self.rot_embs(q, pos_ids).astype(q.dtype)
+            k = self.rot_embs(k, pos_ids).astype(k.dtype)
 
         return q, k, v
